requirements.txt
clip==0.2.0
matplotlib==3.5.3
numpy==1.21.5
Pillow==9.4.0
scikit_learn==1.2.1
torch==1.13.1
torchvision==0.14.1
transformers==4.26.0
